separated out aggregator tests
